fix: dummydata type 분리
